Reduce number of keystrokes :P

+++ mstream.sh
@@ -1,5 +1,9 @@
 #!/bin/bash
 
-# docker build -t mstream:latest .
-mkdir -p output
-docker run --rm -it -v ${PWD}/output:/usr/src/app/output mstream:latest npm run start
+if [[ ${1} == 'build' ]];
+then
+  docker build -t mstream:latest .
+else
+  mkdir -p output
+  docker run --rm -it -v ${PWD}/output:/usr/src/app/output mstream:latest npm run start
+fi
